// huru-react/src/types/wishlistType.ts
import { Product } from "./productType";
import { User } from "./userTypes";

export interface Wishlist {
    id: number;
    user: User;
    products: Product[];
}
export interface WishlistState {
    wishlist: Wishlist | null;
    loading: boolean;
    error: string | null;
}

//payload interfaces for async thunks
export interface AddProductToWishlistPayload {
    wishlistId: number;
    productId: number;
}